pyproject.toml
[project]
name = "fetch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.11",
    "anthropic>=0.50.0",
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "google-ai-generativelanguage>=0.6.15",
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0",
    "google-cloud-aiplatform[tokenization]>=1.91.0",
    "google-cloud-secret-manager>=2.23.3",
    "google-genai>=1.13.0",
    "google-generativeai>=0.8.5",
    "html-to-markdown>=1.3.2",
    "html5lib>=1.1",
    "httplib2>=0.22.0",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "markdownify>=1.1.0",
    "mcp[cli]>=1.7.1",
    "openai>=1.59.4",
    "pyperclip>=1.9.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-xdist>=3.6.1",
    "readabilipy>=0.3.0",
    "ruff>=0.11.9",
    "tiktoken>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
filterwarnings = [
    "ignore::UserWarning:google.auth._default"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests"
]
# Parallel test execution settings
addopts = [
    "--strict-markers",  # Ensure all marks are registered
    "--tb=short",       # Shorter traceback format
    "--dist=worksteal", # Better work distribution for uneven test times
]
